style: split alert tests in separate file
