+++ .secrets.baseline
@@ -91,10 +91,6 @@
       "path": "detect_secrets.filters.allowlist.is_line_allowlisted"
     },
     {
-      "path": "detect_secrets.filters.common.is_baseline_file",
-      "filename": ".secrets.baseline"
-    },
-    {
       "path": "detect_secrets.filters.common.is_ignored_due_to_verification_policies",
       "min_level": 2
     },
@@ -143,23 +139,33 @@
         "filename": "tests/test_openobserve_api.py",
         "hashed_secret": "7bcb636b3be428bee157e3d026bfe5871efe5d24",
         "is_verified": true,
-        "line_number": 492,
+        "is_secret": false,
+        "line_number": 492
+      }
+    ],
+    "tests/test_openobserve_api_alert.py": [
+      {
+        "type": "Secret Keyword",
+        "filename": "tests/test_openobserve_api_alert.py",
+        "hashed_secret": "7bcb636b3be428bee157e3d026bfe5871efe5d24",
+        "is_verified": true,
+        "line_number": 45,
         "is_secret": false
       },
       {
         "type": "Base64 High Entropy String",
-        "filename": "tests/test_openobserve_api.py",
+        "filename": "tests/test_openobserve_api_alert.py",
         "hashed_secret": "ed1768a51d7df2aa6072fe813e54572d6b84cce6",
         "is_verified": true,
-        "line_number": 612,
+        "line_number": 165,
         "is_secret": false
       },
       {
         "type": "Base64 High Entropy String",
-        "filename": "tests/test_openobserve_api.py",
+        "filename": "tests/test_openobserve_api_alert.py",
         "hashed_secret": "d73ecc9b20ee7d6ffe540502bc95505fb782c3b3",
         "is_verified": true,
-        "line_number": 656,
+        "line_number": 209,
         "is_secret": false
       }
     ],
@@ -190,5 +196,5 @@
       }
     ]
   },
-  "generated_at": "2025-06-15T23:37:20Z"
+  "generated_at": "2025-06-22T19:27:28Z"
 }
